fixing warnings, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@26715 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/netbuild/NBNode.cpp
@@ -2303,14 +2303,14 @@ NBNode::buildWalkingAreas(int cornerDetail) {
                     c->valid = false;
                 }
                 c->nextWalkingArea = wa.id;
-                if (c->edges.size() < wa.minPrevCrossingEdges) {
+                if ((int)c->edges.size() < wa.minPrevCrossingEdges) {
                     // if there are multiple crossings, use the shape of the one that crosses fewer edges
                     endCrossingWidth = c->width;
                     endCrossingShape = c->shape;
                     wa.width = MAX2(wa.width, endCrossingWidth);
                     connectsCrossing = true;
                     connectedPoints.push_back(c->shape[-1]);
-                    wa.minPrevCrossingEdges = c->edges.size();
+                    wa.minPrevCrossingEdges = (int)c->edges.size();
                 }
                 if (gDebugFlag1) {
                     std::cout << "    crossing " << c->id << " ends\n";
@@ -2328,14 +2328,14 @@ NBNode::buildWalkingAreas(int cornerDetail) {
                 }
                 c->prevWalkingArea = wa.id;
                 wa.nextCrossings.push_back(c->id);
-                if (c->edges.size() < wa.minNextCrossingEdges) {
+                if ((int)c->edges.size() < wa.minNextCrossingEdges) {
                     // if there are multiple crossings, use the shape of the one that crosses fewer edges
                     startCrossingWidth = c->width;
                     startCrossingShape = c->shape;
                     wa.width = MAX2(wa.width, startCrossingWidth);
                     connectsCrossing = true;
                     connectedPoints.push_back(c->shape[0]);
-                    wa.minNextCrossingEdges = c->edges.size();
+                    wa.minNextCrossingEdges = (int)c->edges.size();
                 }
                 if (gDebugFlag1) {
                     std::cout << "    crossing " << c->id << " starts\n";
